Commit libc headers in this repo
We will use that as the source for establishing the common ground
between libc headers on `aarch64` and `x86_64`.

For the install step, we will use the libc headers committed to this
repo to generate the three directories: `any-macos-any`,
`aarch64-macos-gnu`, and `x86_64-macos-gnu` which will be then installed
into a specified target install dir. Upon installation, we will
proactively remove the three directories from the installation path
so that we avoid obsoletes.